requirements.txt
astroid==2.14.1
click==8.1.3
dill==0.3.6
Flask==2.2.2
Flask-WTF==1.1.1
importlib-metadata==6.0.0
isort==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
lazy-object-proxy==1.9.0
MarkupSafe==2.1.2
mccabe==0.7.0
platformdirs==3.0.0
PyMuPDF==1.21.1
PyPDF2==3.0.1
python-dotenv==0.21.1
tomli==2.0.1
tomlkit==0.11.6
typing_extensions==4.4.0
Werkzeug==2.2.2
wrapt==1.14.1
WTForms==3.0.1
zipp==3.12.1
